+++ vue/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vue-breeze-api",
+  "name": "vue",
   "private": true,
   "version": "0.0.0",
   "type": "module",
