+++ vue/src/router/index.js
@@ -4,10 +4,10 @@ import { useAuthStore } from "../stores/auth";
 //() => import(".. lazyload components...
 const routes = [
     { path: "/", name: "Accueil", component: () => import("../components/Accueil.vue") },
-    { path: "/connecter", name: "Connecter", component: () => import("../components/Connecter.vue") },
     { path: "/creer-compte", name: "CreerCompte", component: () => import("../components/Inscrire.vue") },
     { path: "/profil", name: "Profil", component: () => import("../components/Profil.vue") },
     { path: "/ajouter-bouteille", name: "AjouterBouteille", component: () => import("../components/AjouterBouteille.vue") },
+    { path: '/:pathMatch(.*)*', name: 'not-found', component: () => import("../components/pageNonExistante.vue") },
 ]
 
 const router = createRouter({
